# Conflicts:
#	erp-common/erp-common-core/src/main/java/com/common/core/enums/ApiError.java
#	erp-common/erp-common-message/src/main/java/com/common/message/constant/RocketMqTopic.java
#	erp-common/erp-common-message/src/main/java/com/common/message/enums/RocketMqTagEnum.java
#	erp-model/erp-model-plm/src/main/java/com/erp/model/plm/dto/ProductDetailShowDTO.java
#	erp-server/erp-server-dmp/src/main/java/com/erp/server/dmp/pull/mapper/ProductDetailMapper.java
#	erp-server/erp-server-dmp/src/main/java/com/erp/server/dmp/pull/service/dmp/impl/DmpOrderInfoServiceImpl.java
#	erp-server/erp-server-dmp/src/main/java/com/erp/server/dmp/pull/service/dmp/impl/DmpOrderItemServiceImpl.java
#	erp-server/erp-server-dmp/src/main/resources/mapper/DmpOrderItemMapper.xml
#	erp-server/erp-server-plm/src/main/java/com/erp/server/plm/schedule/PlmJob.java
#	erp-server/erp-server-plm/src/main/java/com/erp/server/plm/service/impl/ProductDetailServiceImpl.java
#	erp-server/erp-server-plm/src/main/java/com/erp/server/plm/service/impl/ProductSaleServiceImpl.java
#	erp-server/erp-server-plm/src/main/resources/mapper/ProductInfoEntityMapper.xml
#	erp-server/erp-server-scm/src/main/java/com/erp/server/scm/listener/SupplierExcelListener.java
#	erp-server/erp-server-scm/src/main/java/com/erp/server/scm/service/PurchasePriceService.java
#	erp-server/erp-server-scm/src/main/java/com/erp/server/scm/service/SupplierGradeService.java
#	erp-server/erp-server-scm/src/main/java/com/erp/server/scm/service/impl/PurchasePriceChangeDetailServiceImpl.java
#	erp-server/erp-server-scm/src/main/java/com/erp/server/scm/service/impl/PurchasePriceServiceImpl.java
#	erp-server/erp-server-scm/src/main/java/com/erp/server/scm/service/impl/SupplierGradeServiceImpl.java
#	erp-server/erp-server-scm/src/main/java/com/erp/server/scm/service/impl/SupplierServiceImpl.java

+++ erp-model/erp-model-bi/src/main/java/com/erp/model/bi/vo/SalesVO.java
@@ -45,7 +45,7 @@ public class SalesVO implements Serializable {
 
 
     /**
-     * 客单假
+     * 客单价
      */
     private BigDecimal perCustomerTransaction=BigDecimal.ZERO;
 
@@ -69,5 +69,5 @@ public class SalesVO implements Serializable {
     /**
      * 销售趋势
      */
-    private List<BigDecimal> salesTrend;
+    private List<Integer> salesTrend;
 }
